// be-camp-cuss/src/common/enums/env-keys.enum.ts
export enum JwtEnvKeys {
  JWT_ACCESS_SECRET = 'JWT_ACCESS_SECRET',
  JWT_REFRESH_SECRET = 'JWT_REFRESH_SECRET',
  JWT_ACCESS_EXPIRES = 'JWT_ACCESS_EXPIRES',
  JWT_REFRESH_EXPIRES = 'JWT_REFRESH_EXPIRES',
}

export enum StorageEnvKeys {
  S3_BUCKET_PUBLIC = 'S3_BUCKET_PUBLIC',
  S3_BUCKET_PRIVATE = 'S3_BUCKET_PRIVATE',
  S3_REGION = 'S3_REGION',
  S3_ENDPOINT = 'S3_ENDPOINT',
  S3_ACCESS_KEY_ID = 'S3_ACCESS_KEY_ID',
  S3_SECRET_ACCESS_KEY = 'S3_SECRET_ACCESS_KEY',
  SUPABASE_PUBLIC_URL = 'SUPABASE_PUBLIC_URL',
  SUPABASE_SERVICE_ROLE_KEY = 'SUPABASE_SERVICE_ROLE_KEY',
}